Update Employee.java
I entered the attributes, made the attribute private according to the question and used getter and setter to access the private fields. Similarly, I have used an integer data type for the date for the more straightforward age calculation. I have also imported the calendar to calculate the age with reference to the current date.